Add VB tests
